Corrected formatting
Output now in either list with tickers or empty list if there isnt any new tickers.
Only tickers with score of 2 or 3 (low float or low float + penny stock) are selected.
seen_tickers.txt file added to store the tickers already analysez. The idea is to periodically reset this file (for example every 24h, since we are not expecting multiple news during the same day)